Some fixes

+++ RRHH MBC Web Application/BeyondThemes.BeyondAdmin/Scripts/Users/_Index/_GroupsList.js
@@ -1,5 +1,5 @@
-﻿var Imtech = {};
-Imtech.Pager = function () {
+﻿var ImtechGroup = {};
+ImtechGroup.Pager = function () {
     this.currentPage = 1;
     this.numPages = function () {
         var numPages = 0;
@@ -14,15 +14,12 @@ Imtech.Pager = function () {
         var count = 0
         this.paragraphs.slice((page - 1) * this.paragraphsPerPage,
             ((page - 1) * this.paragraphsPerPage) + this.paragraphsPerPage).each(function () {
-                console.log($(this).html())
                 html += '<div>' + $(this).html() + '</div>';
                 count += 1;
                 if (count % 3 == 0) {
                     html += '<div class=row></div>'
                 }
             });
-        console.log("hola")
-        console.log(html)
         $(this.pagingContainerPath).html(html);
         renderControls(this.pagingControlsContainer, this.currentPage, this.numPages());
     }
@@ -33,7 +30,7 @@ Imtech.Pager = function () {
         if (currentPage == 1) {
             pagingControls += '<li class="disabled"><a href="#">«</a></li>';
         } else {
-            pagingControls += '<li><a href="#" onclick="pager.showPage(' + previousPage + '); return false;" >«</a></li>';
+            pagingControls += '<li><a href="#" onclick="pagerGroup.showPage(' + previousPage + '); return false;" >«</a></li>';
         }
         // show only +- 3 pages
         var startPage = 1;
@@ -46,7 +43,7 @@ Imtech.Pager = function () {
         }
         for (var i = startPage; i <= finishPage; i++) {
             if (i != currentPage) {
-                pagingControls += '<li><a href="#" onclick="pager.showPage(' + i + '); return false;">' + i + '</a></li>';
+                pagingControls += '<li><a href="#" onclick="pagerGroup.showPage(' + i + '); return false;">' + i + '</a></li>';
             } else {
                 pagingControls += '<li class="active"><a href="#">' + i + ' <span class="sr-only">(current)</span></a></li>';
             }
@@ -54,22 +51,30 @@ Imtech.Pager = function () {
         if (currentPage == numPages) {
             pagingControls += '<li class = disabled><a href="#">»</a></li></ul>';
         } else {
-            pagingControls += '<li><a href="#" onclick="pager.showPage(' + nextPage + '); return false;" >»</a></li></ul>';
+            pagingControls += '<li><a href="#" onclick="pagerGroup.showPage(' + nextPage + '); return false;" >»</a></li></ul>';
         }
         $(container).html(pagingControls);
     }
 }
-var pager = new Imtech.Pager();
+var pagerGroup = new ImtechGroup.Pager();
 var groupContainers;
 $(document).ready(function () {
-    groupContainers = $('#groupList').children();
-    pager.paragraphsPerPage = 9; // set amount elements per page
-    pager.pagingContainer = $('#groupList'); // set of main container
-    pager.paragraphs = $('.groupWidget', pager.pagingContainer); // set of required containers
-    pager.pagingControlsContainer = '#pagingControls';
-    pager.pagingContainerPath = '#groupList';
-    pager.showPage(1);
-    $("[data-toggle='tooltip']").tooltip();
+    var groupListHtml = ""
+    $.ajax({
+        url: "/Users/_GroupList/", type: "GET", dataType: "html",
+        success: function (data) {
+            groupListHtml = data;
+            $('#groupsTab').html(groupListHtml)
+            pagerGroup.pagingControlsContainer = '#pagingControlsGroups'; //string of paging controls
+            pagerGroup.pagingContainerPath = '#groupList'; //string of the main container path
+            groupContainers = $(pagerGroup.pagingContainerPath).children(); //All containers
+            pagerGroup.paragraphsPerPage = 12; // set amount elements per page
+            pagerGroup.pagingContainer = $(pagerGroup.pagingContainerPath); // set of main container
+            pagerGroup.paragraphs = pagerGroup.pagingContainer.children() // set of required containers to search
+            pagerGroup.showPage(1);
+            $("[data-toggle='tooltip']").tooltip();
+        },
+    });
 });
 var delay = (function () {
     var timer = 0;
@@ -92,35 +97,48 @@ function filterGroups(pValue) {
                 groupBox.hide()
             }
         });
-        pager.paragraphs = $('.visible', groupContainers).parent() // set of required containers
-        pager.showPage(1)
+        pagerGroup.paragraphs = $('.visible', groupContainers).parent() // set of required containers
+        pagerGroup.showPage(1)
         $("[data-toggle='tooltip']").tooltip();
     }, 300);
 }
 
 function deleteGroup(group_id, name, self) {
-    $.ajax({
-        url: "/Users/_DeleteGroup/?group_id=" + group_id,
-        type: "DELETE",
-        dataType: "html",
-        traditional: true,
-        contentType: false,
-        success: function (data) {
-            Notify("El grupo '" + name + "' ha sido borrado", 'bottom-right', '5000', 'success', 'fa-edit', true);
-            //var toolbarLink = $(self);
-            //var widget = toolbarLink.parents(".widget").eq(0).parent();
-            var groupBox = $(self).closest('.groupBox')
-            groupBox.hide(0, function () {
-                groupBox.remove();
-                id_groupBox = groupBox.attr('id')
-                groupContainers.find('#' + id_groupBox).remove()
-                pager.paragraphs = $('.visible', groupContainers).parent() // set of required containers
-                pager.showPage(pager.currentPage)
-                $("[data-toggle='tooltip']").tooltip();
+    $.confirm({
+        icon: 'fa fa-warning',
+        keyboardEnabled: true,
+        title: 'Precaución!',
+        content: "El grupo '" + name + "' va a a ser eliminado, seguro que desea eliminarlo?",
+        confirmButton: "Confirmar",
+        cancelButton: "Cancelar",
+        confirm: function () {
+            $.ajax({
+                url: "/Users/_DeleteGroup/?group_id=" + group_id,
+                type: "DELETE",
+                dataType: "html",
+                traditional: true,
+                contentType: false,
+                success: function (data) {
+                    Notify("El grupo '" + name + "' ha sido borrado", 'bottom-right', '5000', 'success', 'fa-edit', true);
+                    var groupBox = $(self).closest('.groupBox')
+                    groupBox.hide(0, function () {
+                        groupBox.remove();
+                        id_groupBox = groupBox.attr('id')
+                        groupContainers.find('#' + id_groupBox).remove()
+                        pagerGroup.paragraphs = $('.visible', groupContainers).parent()
+                        pagerGroup.showPage(pagerGroup.currentPage)
+                        $("[data-toggle='tooltip']").tooltip();
+                    });
+                },
+                error: function () {
+                    Notify("Error, no se puede borrar el grupo '" + name + "'", 'bottom-right', '5000', 'danger', 'fa-edit', true);
+                }
             });
         },
-        error: function () {
-            Notify("Error, no se puede borrar el grupo '" + name + "'", 'bottom-right', '5000', 'danger', 'fa-edit', true);
+        cancel: function () {
         }
     });
+
+
+
 }
